requirements.txt
paho-mqtt~=1.6.1
dash~=2.13.0